langchain-quickbooks-invoice-agent/requirements.txt

langchain>=0.2.14
langchain-openai>=0.1.7
python-dotenv>=1.0.1
fastapi>=0.104.1
uvicorn>=0.24.0
requests>=2.31.0
pydantic>=2.0.0
